refactor: move docs utilities to lib/docs.ts
Extract getAvailableYears, getFirstCourseUrl, and isRootFolder from
lib/source.tsx to a dedicated lib/docs.ts file for better separation.
